requirements.txt
Django==5.2
djangorestframework==3.14.0
dj-database-url==2.1.0
psycopg2-binary==2.9.9
gunicorn==21.2.0
django-cors-headers==4.3.1
requests==2.31.0
beautifulsoup4==4.12.2
whitenoise==6.6.0
openai==1.12.0